+++ Core/include/Acts/Seeding/BinnedSPGroup.hpp
@@ -36,12 +36,14 @@ class BinnedSPGroupIterator {
  private:
   enum INDEX : int { PHI = 0, Z = 1 };
 
- public:
+public:
   // Never take ownerships
   BinnedSPGroupIterator(BinnedSPGroup<external_spacepoint_t>&& group,
-                        std::size_t) = delete;
+                        std::array<std::size_t, 2> index,
+			std::array<std::vector<std::size_t>, 2> navigation) = delete;
   BinnedSPGroupIterator(BinnedSPGroup<external_spacepoint_t>& group,
-                        std::size_t index);
+                        std::array<std::size_t, 2> index,
+			std::array<std::vector<std::size_t>, 2> navigation);
 
   BinnedSPGroupIterator(const BinnedSPGroupIterator&) = delete;
   BinnedSPGroupIterator& operator=(const BinnedSPGroupIterator&) = delete;
@@ -66,10 +68,10 @@ class BinnedSPGroupIterator {
  private:
   // The group, it contains the grid and the bin finders
   Acts::detail::RefHolder<BinnedSPGroup<external_spacepoint_t>> m_group;
-  // Max Local Bins - limits of the grid
-  std::array<std::size_t, 2> m_max_localBins;
-  // Current Local Bins
-  std::array<std::size_t, 2> m_current_localBins{0, 0};
+  // Current grid iterator
+  typename Acts::SpacePointGrid<external_spacepoint_t>::local_iterator_t m_gridItr;
+  // End iterator
+  typename Acts::SpacePointGrid<external_spacepoint_t>::local_iterator_t m_gridItrEnd;
 };
 
 /// @c BinnedSPGroup Provides access to begin and end BinnedSPGroupIterator
@@ -81,6 +83,8 @@ class BinnedSPGroup {
   friend BinnedSPGroupIterator<external_spacepoint_t>;
 #endif
 
+  enum INDEX : int { PHI = 0, Z = 1 };
+  
  public:
   BinnedSPGroup() = delete;
 
@@ -119,17 +123,17 @@ class BinnedSPGroup {
 
  private:
   // grid with ownership of all InternalSpacePoint
-  std::unique_ptr<Acts::SpacePointGrid<external_spacepoint_t>> m_grid;
+  std::unique_ptr<Acts::SpacePointGrid<external_spacepoint_t>> m_grid {nullptr};
 
   // BinFinder must return std::vector<Acts::Seeding::Bin> with content of
   // each bin sorted in r (ascending)
-  std::shared_ptr<const BinFinder<external_spacepoint_t>> m_topBinFinder;
-  std::shared_ptr<const BinFinder<external_spacepoint_t>> m_bottomBinFinder;
+  std::shared_ptr<const BinFinder<external_spacepoint_t>> m_topBinFinder {nullptr};
+  std::shared_ptr<const BinFinder<external_spacepoint_t>> m_bottomBinFinder {nullptr};
 
   // Order of z bins to loop over when searching for SPs
-  std::vector<std::size_t> m_bins;
+  std::array<std::vector<std::size_t>, 2> m_bins {};
   // Number of Z bins to skip the search for middle SP
-  std::size_t m_skipZMiddleBin;
+  std::size_t m_skipZMiddleBin {0ul};
 };
 
 }  // namespace Acts
