+++ Core/include/Acts/Seeding/BinnedSPGroup.ipp
@@ -1,3 +1,4 @@
+// -*- C++ -*-
 // This file is part of the Acts project.
 //
 // Copyright (C) 2023 CERN for the benefit of the Acts project
@@ -12,45 +13,33 @@
 
 template <typename external_spacepoint_t>
 Acts::BinnedSPGroupIterator<external_spacepoint_t>::BinnedSPGroupIterator(
-    Acts::BinnedSPGroup<external_spacepoint_t>& group, std::size_t index)
-    : m_group(group), m_max_localBins(m_group->m_grid->numLocalBins()) {
-  m_max_localBins[INDEX::PHI] += 1;
-  m_current_localBins[INDEX::Z] = m_group->skipZMiddleBin();
-  if (index == m_group->m_grid->size()) {
-    m_current_localBins = m_max_localBins;
-  } else {
-    // Go to the next not-empty bin
-    findNotEmptyBin();
-  }
+									  Acts::BinnedSPGroup<external_spacepoint_t>& group,
+									  std::array<std::size_t, 2> index,
+									  std::array<std::vector<std::size_t>, 2> navigation)
+    : m_group(group),
+      m_gridItr(*group.m_grid.get(), std::move(index), navigation),
+      m_gridItrEnd(*group.m_grid.get(), group.m_grid->numLocalBins(), std::move(navigation))
+{
+  findNotEmptyBin(); 
 }
 
 template <typename external_spacepoint_t>
 inline Acts::BinnedSPGroupIterator<external_spacepoint_t>&
 Acts::BinnedSPGroupIterator<external_spacepoint_t>::operator++() {
-  // Increase the position by one
-  // if we were on the edge, go up one phi bin and reset z bin
-  if (++m_current_localBins[INDEX::Z] == m_max_localBins[INDEX::Z]) {
-    ++m_current_localBins[INDEX::PHI];
-    m_current_localBins[INDEX::Z] = m_group->skipZMiddleBin();
-  }
-
-  // Get the next not-empty bin in the grid
+  ++m_gridItr;
   findNotEmptyBin();
   return *this;
 }
 
 template <typename external_spacepoint_t>
 inline bool Acts::BinnedSPGroupIterator<external_spacepoint_t>::operator==(
-    const Acts::BinnedSPGroupIterator<external_spacepoint_t>& other) const {
-  return m_group.ptr == other.m_group.ptr &&
-         m_current_localBins[INDEX::PHI] ==
-             other.m_current_localBins[INDEX::PHI] &&
-         m_current_localBins[INDEX::Z] == other.m_current_localBins[INDEX::Z];
+									   const Acts::BinnedSPGroupIterator<external_spacepoint_t>& other) const {
+  return m_group.ptr == other.m_group.ptr && m_gridItr == other.m_gridItr;
 }
 
 template <typename external_spacepoint_t>
 inline bool Acts::BinnedSPGroupIterator<external_spacepoint_t>::operator!=(
-    const Acts::BinnedSPGroupIterator<external_spacepoint_t>& other) const {
+										   const Acts::BinnedSPGroupIterator<external_spacepoint_t>& other) const {
   return !(*this == other);
 }
 
@@ -59,19 +48,18 @@ std::tuple<boost::container::small_vector<std::size_t, 9>, std::size_t,
            boost::container::small_vector<std::size_t, 9>>
 Acts::BinnedSPGroupIterator<external_spacepoint_t>::operator*() const {
   // Global Index
-  std::size_t global_index = m_group->m_grid->globalBinFromLocalBins(
-      {m_current_localBins[INDEX::PHI],
-       m_group->m_bins[m_current_localBins[INDEX::Z]]});
-
+  std::array<std::size_t, 2> localPosition = m_gridItr.localPosition();
+  std::size_t global_index = m_group->m_grid->globalBinFromLocalBins(localPosition);
+  
   boost::container::small_vector<std::size_t, 9> bottoms =
       m_group->m_bottomBinFinder->findBins(
-          m_current_localBins[INDEX::PHI],
-          m_group->m_bins[m_current_localBins[INDEX::Z]],
+	  localPosition[INDEX::PHI],
+	  localPosition[INDEX::Z],
           m_group->m_grid.get());
   boost::container::small_vector<std::size_t, 9> tops =
       m_group->m_topBinFinder->findBins(
-          m_current_localBins[INDEX::PHI],
-          m_group->m_bins[m_current_localBins[INDEX::Z]],
+          localPosition[INDEX::PHI],
+          localPosition[INDEX::Z],
           m_group->m_grid.get());
 
   // GCC12+ in Release throws an overread warning here due to the move.
@@ -89,37 +77,13 @@ Acts::BinnedSPGroupIterator<external_spacepoint_t>::operator*() const {
 template <typename external_spacepoint_t>
 inline void
 Acts::BinnedSPGroupIterator<external_spacepoint_t>::findNotEmptyBin() {
+  if (m_gridItr == m_gridItrEnd) return;
   // Iterate on the grid till we find a not-empty bin
   // We start from the current bin configuration and move forward
-
-  for (std::size_t phiBin(m_current_localBins[INDEX::PHI]);
-       phiBin < m_max_localBins[INDEX::PHI]; ++phiBin) {
-    // 0 is the underflow - skip
-    if (phiBin == 0) {
-      continue;
-    }
-
-    for (std::size_t zBin(m_current_localBins[INDEX::Z]);
-         zBin < m_max_localBins[INDEX::Z]; ++zBin) {
-      std::size_t zBinIndex = m_group->m_bins[zBin];
-      std::size_t index =
-          m_group->m_grid->globalBinFromLocalBins({phiBin, zBinIndex});
-      // Check if there are entries in this bin
-      if (m_group->m_grid->at(index).empty()) {
-        continue;
-      }
-
-      // Set the new current bins
-      m_current_localBins[INDEX::PHI] = phiBin;
-      m_current_localBins[INDEX::Z] = zBin;
-      return;
-    }
-    // Reset z-index
-    m_current_localBins[INDEX::Z] = m_group->skipZMiddleBin();
+  std::size_t dimCollection = (*m_gridItr).size();
+  while (dimCollection == 0ul and ++m_gridItr != m_gridItrEnd) {
+    dimCollection = (*m_gridItr).size();
   }
-
-  // Could find nothing ... setting this to end()
-  m_current_localBins = m_max_localBins;
 }
 
 // Binned SP Group
@@ -231,13 +195,13 @@ Acts::BinnedSPGroup<external_spacepoint_t>::BinnedSPGroup(
   m_topBinFinder = tBinFinder;
 
   m_skipZMiddleBin = config.skipZMiddleBinSearch;
-  m_bins = config.zBinsCustomLooping;
-  if (m_bins.empty()) {
+  m_bins[INDEX::PHI].resize(m_grid->numLocalBins()[0]);
+  std::iota(m_bins[INDEX::PHI].begin(), m_bins[INDEX::PHI].end(), 1);
+  m_bins[INDEX::Z] = config.zBinsCustomLooping;
+  if (m_bins[INDEX::Z].empty()) {
     std::size_t nZbins = m_grid->numLocalBins()[1];
-    m_bins.reserve(nZbins);
-    for (std::size_t i(0); i < nZbins; ++i) {
-      m_bins.push_back(i + 1);
-    }
+    m_bins[INDEX::Z].resize(nZbins);
+    std::iota(m_bins[INDEX::Z].begin(), m_bins[INDEX::Z].end(), 1);
   }
 }
 
@@ -249,11 +213,11 @@ inline std::size_t Acts::BinnedSPGroup<external_spacepoint_t>::size() const {
 template <typename external_spacepoint_t>
 inline Acts::BinnedSPGroupIterator<external_spacepoint_t>
 Acts::BinnedSPGroup<external_spacepoint_t>::begin() {
-  return {*this, 0};
+  return {*this, {0ul, 0ul}, m_bins};
 }
 
 template <typename external_spacepoint_t>
 inline Acts::BinnedSPGroupIterator<external_spacepoint_t>
 Acts::BinnedSPGroup<external_spacepoint_t>::end() {
-  return {*this, m_grid->size()};
+  return {*this, m_grid->numLocalBins(), m_bins};
 }
